+++ dsmr_backend/mixins.py
@@ -35,6 +35,7 @@ class InfiniteManagementCommandMixin(object):
 
     def handle(self, **options):
         """ Called by Django to run command. We relay to run() ourselves and keep it running. """
+        self._check_logger_level()
         self._write_pid_file()
         self.data = self.initialize()
 
@@ -123,6 +124,14 @@ class InfiniteManagementCommandMixin(object):
             os.unlink(self._pid_file)
         except IOError:
             pass
+
+    def _check_logger_level(self):
+        # This will result in only logging errors, so make sure to clear that up.
+        if logger.getEffectiveLevel() > logging.INFO:
+            print(
+                'The current logging level only logs warnings and errors, to reduce I/O. More information can be '
+                'found here: https://dsmr-reader.readthedocs.io/en/latest/troubleshooting.html#logging'
+            )
 
 
 class ReadOnlyAdminModel(admin.ModelAdmin):