+++ dsmr_datalogger/management/commands/dsmr_datalogger.py
@@ -1,3 +1,5 @@
+import logging
+
 from django.core.management.base import BaseCommand, CommandError
 from django.utils.translation import ugettext as _
 from django.conf import settings
@@ -6,6 +8,9 @@ from dsmr_backend.mixins import InfiniteManagementCommandMixin
 from dsmr_datalogger.models.settings import DataloggerSettings
 from dsmr_datalogger.exceptions import InvalidTelegramError
 import dsmr_datalogger.services
+
+
+logger = logging.getLogger('commands')
 
 
 class Command(InfiniteManagementCommandMixin, BaseCommand):
@@ -22,9 +27,7 @@ class Command(InfiniteManagementCommandMixin, BaseCommand):
             raise CommandError("Datalogger tracking is DISABLED!")
 
         telegram = dsmr_datalogger.services.read_telegram()
-
-        # Reflect output to STDOUT for logging and convenience.
-        self.stdout.write(telegram)
+        logger.info("\n%s", telegram)
 
         try:
             dsmr_datalogger.services.telegram_to_reading(data=telegram)